Preparing for version 3.0
After discussing with Ian Mackenzie, Ilias Van Peer and Folkert de Vries in elm-slack, I reached the conclusion that this API (which had initially been inspired by elm-core's List) was exposing different (and confusing) ways of interacting with a sliding list. This commit greatly reduces the surface for this confusion. We now have a `fromList`, an `insert` and `insertAll` and they all behave the same way: when you insert something into a sliding list, this item is inserted to the end of the sliding list, and items will drop from the beginning of the sliding list if necessary. This is not a performance driven implementation, so YMMV. In the tests I have also included some usage examples.